Dev (#15)
* Update docker-compose file

Change container names to reflect the new bot name

* Set playing game to "!!help"

Set the presence status to "!!help"" on startup

* Whoops

Change the name from rpBot to CinnamonRP throughout the codebase

* Add invite command

Added "!!invite" command that generates OAuth invite URL for current bot instance

* Clean up code

Use non-deprecated methods and clean up code

* Fixed Sample-config; improved invite command (now info command)

* Created Location list and location creation

* Added location searching

* Added quote detection

* Added location description

* Added auto reconnect

* Renamed webhook stuff

* Added reference management

* Added deleteing locations. Locations are complete

* Updated description

* Add database backup scripts

Added database backup scripts to dump and restore the data from the Docker mongo container

* Add basic info message

Add a basic information message and invite link to the official support server

* Add more command aliases

Rework the location commands and subcommands, added common synonyms as aliases.

Change the behaviour of command erros to echo the original parameter(s).

* Update TODO

Rename stuff and check off completed items

* Added react timeout

* Added some cleanup, increased time

* Added population as to not re-search database for files

* Log more info to console on startup

Print ID of bot user and number of guilds to console on bot startup

* Add server tracker

Add server count in presence status when server count is greater than 1

* Add character command

Add skeleton of character command. TODO: message proxying

* Added character population

* Added user testing

* Added user testing

* I was an idiot

* Added character owner to character list

* Added user checking on reactions

* Prevent crashing on timeout after character deleted

+++ bot/commands/character.js
@@ -36,7 +36,7 @@ Removes a reference from \`<character>\``,
       response.description = ""
       for (var index = 0; index < charactersList.length; index++) {
         var character = charactersList[index]
-        response.description += `\n**${character.name}** \`(${character._id})\` `
+        response.description += `\n**${character.name}** \`(${character._id})\`<@${character.owner}>`
       }
       return msg.channel.send(response)
     }
@@ -163,6 +163,7 @@ Removes a reference from \`<character>\``,
 							}
 							setTimeout((response, reactions) => {
 								reactions.findById(response.id, (err, doc) => {
+                  if(doc == null) return
 									response.clearReactions()
 									reactions.deleteOne({_id: doc._id}, err =>{
 					          if(err) return console.log(err)
